feature/1: use more 'correct' circuit breaker

+++ v4/src/apiserver/core/ports/reservation/reservation.go
@@ -5,7 +5,7 @@ import "context"
 type Config struct {
 	Host     string
 	Port     string
-	MaxFails uint64 `mapstructure:"max_fails"`
+	MaxFails uint32 `mapstructure:"max_fails"`
 }
 
 type Client interface {
